pyproject.toml
[project]
name = "yahoo-ai-gm"
version = "0.1.0"
description = "AI GM assistant for Yahoo Fantasy Baseball (H2H 5x5)."
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
  "pydantic>=2.6.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6.0",
  "pytest>=8.0.0",
]

service = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
]

[tool.ruff]
line-length = 100
